+++ src/reportlab/__init__.py
@@ -1,9 +1,9 @@
 #Copyright ReportLab Europe Ltd. 2000-2023
 #see license.txt for license details
 __doc__="""The Reportlab PDF generation library."""
-Version = "4.0.5"
+Version = "4.0.6"
 __version__=Version
-__date__='20230829'
+__date__='20231010'
 
 import sys, os